Fix issue where different delimiter characters are used
The issue caused anything past the second delimiter to be ignored, since it only looked at the first two parts. The new implementation allows for : and = to be used as part of the value. Note that this not solve the inability to use , or ;. If these are required the user will need to provide a json input