Use human readable names for templates (#162)
Breaking: all locations where a template ID were expected, now expect a
template name. That means the workspace id MUST be in context whenever a
user wants to interact with a template, the arguments in all
template-related commands have been modified as such.

As an extra change, template list now show the beginning of their description
instead on an ID.

Closes: FP-2316